@@ -9,13 +9,14 @@
         }
     ],
     "require": {
-        "php": "^7.2",
-        "monolog/monolog": "^1.23 || ^2.0",
-        "psr/log": "^1.0"
+        "php": "^8",
+        "monolog/monolog": "^2.0",
+        "psr/log": "^2",
+        "thecodingmachine/safe": "^1.3.3",
+        "wyrihaximus/psr-3-utilities": "^2.1"
     },
     "require-dev": {
-        "wyrihaximus/cs-fixer-config": "^1.0",
-        "wyrihaximus/test-utilities": "^1.1"
+        "wyrihaximus/test-utilities": "^4.0.0"
     },
     "autoload": {
         "psr-4": {
@@ -28,62 +29,23 @@
         }
     },
     "config": {
-        "sort-packages": true,
+        "allow-plugins": {
+            "infection/extension-installer": true,
+            "dealerdirect/phpcodesniffer-composer-installer": true,
+            "icanhazstring/composer-unused": true,
+            "ergebnis/composer-normalize": true
+        },
         "platform": {
-            "php": "7.2"
-        }
+            "php": "8.0.12"
+        },
+        "sort-packages": true
     },
     "scripts": {
-        "ensure-installed": "composer install --ansi -n -q",
-        "cs": [
-            "@ensure-installed",
-            "php-cs-fixer fix --config=.php_cs --ansi --dry-run --diff --verbose --allow-risky=yes --show-progress=estimating"
+        "post-install-cmd": [
+            "composer normalize"
         ],
-        "cs-fix": [
-            "@ensure-installed",
-            "php-cs-fixer fix --config=.php_cs --ansi --verbose --allow-risky=yes --show-progress=estimating"
-        ],
-        "unit": [
-            "@ensure-installed",
-            "phpunit --colors=always -c phpunit.xml.dist"
-        ],
-        "unit-coverage": [
-            "@ensure-installed",
-            "phpunit --colors=always -c phpunit.xml.dist --coverage-text --coverage-html covHtml --coverage-clover ./build/logs/clover.xml"
-        ],
-        "lint-php": [
-            "@ensure-installed",
-            "parallel-lint --exclude vendor ."
-        ],
-        "qa-all": [
-            "@lint-php",
-            "@cs",
-            "@unit"
-        ],
-        "qa-all-coverage": [
-            "@lint-php",
-            "@cs",
-            "@unit-coverage"
-        ],
-        "qa-windows": [
-            "@lint-php",
-            "@cs",
-            "@unit"
-        ],
-        "qa-ci": [
-            "@unit"
-        ],
-        "qa-ci-extended": [
-            "@qa-all-coverage"
-        ],
-        "qa-ci-windows": [
-            "@qa-windows"
-        ],
-        "qa-contrib": [
-            "@qa-all"
-        ],
-        "ci-coverage": [
-            "if [ -f ./build/logs/clover.xml ]; then wget https://scrutinizer-ci.com/ocular.phar && php ocular.phar code-coverage:upload --format=php-clover ./build/logs/clover.xml; fi"
+        "post-update-cmd": [
+            "composer normalize"
         ]
     }
 }
